Add group handling within maps
Allow redirection into a named group.

Allow recursion into all the members of a Node::Group.

Convert KeyValue to be a Node; this may permit nonsensical trees because
a naked KeyValue simply may not appear in many places.  But since a
group name may occur anywhere a key-value pair is expected, it seems
easiest.