delete Validate trait
I don't think this is the right way to implement flexibility across
serializated data types.  cbor.rs is already hundreds of lines of
code, we don't want to cut-and-paste that much code for each new type.

This removes the need for interior mutability in WorkingMap and
WorkingArray (which only existed to satisfy trait restrictions).

This also removes the need for lifetime annotations on the Context
parameter.  So this is a big win for readability.