add json support
JSON support turns out to be very easy.  Except for some vagueness about
integer type, it's easy to convert serde_json Value to our Value format.

Also add feature flags for serde_json and serde_cbor, so it's possible
to not build both if you don't need them.

The json tests are pretty-much cut-and-paste from the cbor tests, which
seems like a long-term maintenance hassle.  Perhaps in the future I'll
find a way to share most of the test case logic.