fix identifier parsing bug
RFC8610 2.2.2.1 points out that "min..max" is not a range, but an
identifier.  This is because '.' is a valid ident character, as long as
it doesn't come at the beginning or end of the identifier.

It turns out I hadn't implemented identifier parsing correctly, and
"a..b" was rejected.  This change fixes the parser to follow the rfc
more closely.