implement array group backtracking
If an array contains an occurrence that contains a group, we need to be
able to back out any elements consumed if the entire group fails to
match.

The same thing should happen if there is a choice between two groups.
My conclusion is that snapshot/backtracking is only needed upon entering
and leaving a group.