fix bug handling group choices nested within arrays
Node::Choice needs special handling so we don't lose our array context.
As long as we remember we're inside an array we will correctly snapshot
as we enter each group and either commit our changes or rewind when that
group succeeds or fails.

Add a few more unit tests to enforce proper group choice behavior.