implement map group backtracking
This change is analogous to the change from two commits past, where we
added backtracking to arrays.  If a map contains an occurrence or choice
that contains a group, we need to be able to back out any elements
consumed if the entire group fails to match.

There is a fair amount of duplicated code between the map and array
handling, but there are just enough differences that I'm not sure if
trying to combine them would make the code any more readable.