honor cut/non-cut semantics in validation
When matching map key-value pairs, there is special syntax to indicate
what should happen when a key matches but the value doesn't match.

With "cut" semantics, a key match causes that particular key-value pair
to be "used up"; it may not match against any other CDDL key-value
pairs.

Without "cut" semantics, a partial key-value match may be tried against
other CDDL key-value pairs.

Add a new error called MapCut; it's the same thing as a Mismatch, except
that it signals to our caller that an Occurrence should fail, even if it
was otherwise happy with the number of successes it encountered.

Make changes to the way we interact with WorkingMap: we no longer remove
things on a key match; we only remove them once the value has matched as
well.

Disable a unit test that got caught up in the difference between serde
serialization details and map-cut semantics.

Ensure that "fatal" errors (anything except a mismatch) will trigger a
validation failure and not get accidentally swallowed by a choice or
occurrence.