remove a poor unit test
I don't think it makes much sense to try to test a standalone group with
an occurrence in it.  If it turns out this is useful, I can bring it
back later, but for now, I don't plan on fixing it so I'd like it to go
away so all the unit tests in the code base actually work.