Add a way to preserve the the string associated with a rule
When parsing (and flattening) a CDDL input, add a few functions to allow
the caller to not only get the parsed (flattened) tree representing the
meaning of a rule, but also a String containing the original text that
defined the rule itself.

I confess this probably isn't generally useful, but it's something I
want for my own purposes.