Be more careful using 'as' for integer conversion
We don't want integer truncation on 32-bit platforms, so don't do
"u64 as usize"; TryInto or TryFrom is what we want.

Enable one of the pedantic clippy lints to help catch these in the
future.  It's not likely this library will use deliberate truncation
anywhere, but we can deal with that if/when it happens.