implement flattening of generic arguments
This adds generic types arguments to the IVT, and populates the new
fields based on parser data.

Add a new error variant, used when we detect an inappropriate use of
generics (e.g. a generic parameter used on a prelude type).

Add a bunch of new test helper functions, so the tests can be more
precise (less matching against debug strings).  Add PartialEq to IVT
types so that we can assert equality of flatten results.