implement the "choiceify" operator ("&")
This operator turns a group into a choice. Its operation in a map
context is a bit under-specified in the RFC, so I have chosen to
interpret the spec as though keys are ignored everywhere.

There are a few test cases for maps that don't work as expected-- there
are test cases that are disabled until I work out what's wrong.