make parser errors allocate less
The parser will generate lots of ephemeral errors while exploring the
many different subparsers, so having parse errors allocate heap memory
is a bad idea.

Instead, use a Cow<str> to store the error context, and convert it to an
owned form if the error is permanent and we're exiting the parser.